refactor: extract configuration and README generation logic
Extracted configuration management and README generation logic into
separate modules for better organization and maintainability.
