+++ src/readmai/generator/__init__.py
@@ -1,0 +1,1 @@
+# Generator package initialization
