+++ src/readmai/scanner/__init__.py
@@ -1,0 +1,1 @@
+# Project scanner package initialization
